Former-commit-id: b84a884befd14855baa2d349e664c24de25219b6

+++ ifs-data-service/src/test/java/org/innovateuk/ifs/application/controller/ApplicationAssessmentSummaryControllerIntegrationTest.java
@@ -66,7 +66,7 @@ public class ApplicationAssessmentSummaryControllerIntegrationTest extends BaseC
 
         ApplicationAssessmentSummaryResource applicationAssessmentSummary = serviceResult.getSuccessObjectOrThrowException();
 
-        assertEquals(application.getId(), applicationAssessmentSummary.getId());
+        assertEquals(application.getId().longValue(), applicationAssessmentSummary.getId());
         assertEquals(application.getName(), applicationAssessmentSummary.getName());
         assertEquals(competition.getId(), applicationAssessmentSummary.getCompetitionId());
         assertEquals(competition.getName(), applicationAssessmentSummary.getCompetitionName());
